// src/interfaces/RailwayStopsInterface.ts
export default interface RailwayStopsInterface {
    stop_id: string;
    stop_code: string;
    stop_name: string;
    stop_desc: string;
    stop_lat: string,
    stop_lon: string,
    zone_id: string;
    stop_url: string;
    location_type: string;
    parent_station: string;
    platform_code: string;
}